+++ src/android/Diagnostic_External_Storage.java
@@ -25,7 +25,7 @@ package cordova.plugins;
 import android.os.Build;
 import android.os.Environment;
 import android.os.StatFs;
-import android.support.v4.os.EnvironmentCompat;
+import androidx.core.os.EnvironmentCompat;
 import android.util.Log;
 
 import org.apache.cordova.CallbackContext;