+++ src/android/Diagnostic_Notifications.java
@@ -32,7 +32,7 @@ import org.apache.cordova.CordovaWebView;
 import org.json.JSONArray;
 import org.json.JSONException;
 
-import android.support.v4.app.NotificationManagerCompat;
+import androidx.core.app.NotificationManagerCompat;
 
 /**
  * Diagnostic plugin implementation for Android