+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+* Support new location accuracy authorization API on iOS 14+.
+    * Resolves [#402](https://github.com/dpa99c/cordova-diagnostic-plugin/issues/402).
+* *BREAKING CHANGE:* (Android) Migrate to AndroidX from Android Support Library
+    * Minimum Cordova Android platform version is now `cordova-android@9.0.0`
+    * Minimum supported Android version is now Android 5.1 / API 22    
+
 **v5.0.2**
 * (Android) Add support for background location permission on Android 10 / API 29. Resolves [#398](https://github.com/dpa99c/cordova-diagnostic-plugin/issues/398).
 * (Doc) Add doc about `locationAuthorizationMode` constants